Fixed duplicate code; Removed index from import paths
