+++ src/extensions/RecipeIngredients/RecipeIngredients.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { locations } from 'contentful-ui-extensions-sdk';
 import IngredientsList from './IngredientsList';
 import './RecipeIngredients.scss';
-import { IngredientDialog } from './dialogs/index';
+import { IngredientDialog } from './dialogs';
 
 const RecipeIngredients = ({ sdk }) => {
   return !sdk.location.is(locations.LOCATION_DIALOG)
