+++ src/extensions/RecipeIngredients/dialogs/IngredientDialog.js
@@ -1,7 +1,8 @@
 import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
 import { Form, FieldGroup } from '@contentful/forma-36-react-components';
-import { withLabel, getTextInput, getSelect, getButton } from '../helpers/index';
+import { withLabel, getTextInput, getSelect } from '../helpers';
+import { getButton } from '../../../shared/helpers'; 
 
 const imperialUnits = ['Teaspoon','Teaspoons','Tablespoon','Tablespoons','Cup','Cups','Ounce','Ounces','Pound','Pounds','Gram','Grams','Pinch','Each','As needed','To serve','To Taste','Bunch','Bunches','Can','Cans','Clove','Cloves','Leaf','Leaves','Package','Packages','Recipe','Recipes','Rib','Ribs','Slice','Slices','Wedge','Wedges'];
 const metricUnits = ['Milliliter','Milliliters','Liter','Liters','Ounce','Ounces','Gram','Grams','Pinch','Each','As needed','To serve','To Taste','Bunch','Bunches','Can','Cans','Clove','Cloves','Leaf','Leaves','Package','Packages','Recipe','Recipes','Rib','Ribs','Slice','Slices','Wedge','Wedges'];
