+++ src/shared/helpers/index.js
@@ -1,0 +1,3 @@
+import { getButton, getIconButton } from './formControl';
+
+export { getButton, getIconButton };
